Kleine Aenderung damit auch static geht.

+++ tuxbox/neutrino/daemons/sectionsd/sectionsd.cpp
@@ -23,6 +23,9 @@
 //    Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 //
 //  $Log$
+//  Revision 1.22  2001/07/17 14:15:52  fnbrd
+//  Kleine Aenderung damit auch static geht.
+//
 //  Revision 1.21  2001/07/17 13:14:59  fnbrd
 //  Noch ne Verbesserung in Bezug auf alte Events.
 //
@@ -1009,7 +1012,7 @@ const unsigned timeoutInSeconds=2;
       for(SIevents::iterator e=eit.events().begin(); e!=eit.events().end(); e++)
         if(e->times.size()>0) {
 	  if(e->times.begin()->startzeit<zeit+(long)HOURS_TO_CACHE*60L*60L &&
-	    e->times.begin()->startzeit+e->times.begin()->dauer>zeit-(long)OLD_EVENTS_ARE*60L
+	    e->times.begin()->startzeit+(long)e->times.begin()->dauer>zeit-(long)OLD_EVENTS_ARE*60L
 	  ) {
             pthread_mutex_lock(&eventsLock);
             events.insert(*e);
